+++ extensions/amp-story-auto-ads/0.1/amp-story-auto-ads.css
