+++ extensions/amp-story-auto-ads/0.1/test/test-amp-story-auto-ads.js
