+++ extensions/amp-story/0.1/amp-story.css
@@ -16,7 +16,7 @@
 
 @import './amp-story-desktop.css';
 @import './amp-story-user-overridable.css';
-@import './amp-story-ads.css';
+@import './amp-story-hint.css';
 @import './amp-story-share.css';
 
 /** Common */
