+++ extensions/amp-story/0.1/amp-story.js
@@ -24,7 +24,6 @@
  * </amp-story>
  * </code>
  */
-import './amp-story-auto-ads';
 import './amp-story-cta-layer';
 import './amp-story-grid-layer';
 import './amp-story-page';
